Nearly finished add contract, just need to implement status

+++ src/main/java/contracts/repository/AccountRepository.java
@@ -3,6 +3,6 @@ package contracts.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import contracts.domain.user;
 
-public interface AccountRepository extends JpaRepository<user, Long> {
+public interface AccountRepository extends JpaRepository<user, Integer> {
 
 }
